requirement.txt
asgiref==3.2.5
beautifulsoup4==4.8.2
certifi==2019.11.28
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
Django==3.0.4
django-coreapi==2.3.0
django-filter==2.2.0
django-ranged-response==0.2.0
django-simple-captcha==0.5.12
djangorestframework==3.11.0
djangorestframework-simplejwt==4.4.0
dnspython==1.16.0
future==0.18.2
gevent==1.4.0
greenlet==0.4.15
idna==2.9
itypes==1.1.0
Jinja2==2.11.1
Markdown==3.2.1
MarkupSafe==1.1.1
nmap==0.0.1
Pillow==7.1.1
PyJWT==1.7.1
PyMySQL==0.9.3
python-whois==0.7.2
pytz==2019.3
requests==2.23.0
sentry-sdk==0.14.2
six==1.14.0
soupsieve==2.0
sqlparse==0.3.1
uritemplate==3.0.1
urllib3==1.25.8
